refactor(pxar-mount): reorganize test files by category
