--- internal/pxarmount/bugfix_test.go
@@ -1,446 +1,0 @@
-package pxarmount
-
-import (
-	"sync"
-	"syscall"
-	"testing"
-
-	"github.com/hanwen/go-fuse/v2/fuse"
-)
-
-// Bug A: Readlink, ListXAttr, GetXAttr call readEntryForNode without
-// holding readerMu. readEntryForNode -> ReadEntryForNode -> reader.ReadEntryAt
-// requires readerMu. Read and ReadDirRaw hold readerMu but Readlink and
-// the xattr methods do not, creating a data race on the shared archive reader.
-//
-// This test confirms the race by running Readlink concurrently with a reader
-// that's being used via Read (which acquires readerMu).
-
-func TestReadlink_HoldsReaderMu(t *testing.T) {
-	// Verifies Readlink calls readEntryForNode under readerMu.
-	// Root inode path uses a synthetic entry (no reader needed),
-	// but the method still acquires readerMu. We verify no panic.
-	fs, _, cleanup := newTestPassthroughFS(t)
-	defer cleanup()
-
-	fs.pxar.mu.Lock()
-	fs.pxar.nodes[RootInode] = node{
-		inode:     RootInode,
-		parent:    RootInode,
-		mode:      uint64(syscall.S_IFLNK | 0o777),
-		isSymlink: true,
-	}
-	fs.pxar.mu.Unlock()
-
-	// Root entry is not actually a symlink so Readlink returns EINVAL.
-	// The point is that readerMu is properly acquired without deadlock.
-	_, status := fs.pxar.Readlink(nil, &fuse.InHeader{NodeId: RootInode})
-	// Root returns OK (synthetic entry), but LinkTarget is empty.
-	// The important thing: no panic or race detected.
-	_ = status
-}
-
-// Bug B: Concurrent HotSwap + Read races on readerAt.
-// HotSwap writes fs.readerAt under mu, while Read reads it under mu.RLock.
-// After the fix, no race should be detected.
-
-func TestHotSwap_ReaderAtRace(t *testing.T) {
-	pxarFS := &PxarFS{
-		nodes: make(map[uint64]node),
-	}
-	pxarFS.nodes[RootInode] = node{
-		inode:  RootInode,
-		parent: RootInode,
-		mode:   uint64(syscall.S_IFDIR | 0o755),
-		isDir:  true,
-		refs:   1,
-	}
-
-	var wg sync.WaitGroup
-	const rounds = 100
-
-	// Simulate concurrent reads accessing readerAt (using proper lock)
-	wg.Go(func() {
-		for range rounds {
-			pxarFS.mu.RLock()
-			_ = pxarFS.readerAt
-			pxarFS.mu.RUnlock()
-		}
-	})
-
-	// Simulate HotSwap writing readerAt (using proper lock)
-	wg.Go(func() {
-		for range rounds {
-			pxarFS.mu.Lock()
-			pxarFS.readerAt = nil
-			pxarFS.mu.Unlock()
-		}
-	})
-
-	wg.Wait()
-	// With -race this should NOT flag any data race
-}
-
-// Bug C: releaseDirEntries pools the same underlying array on Put/Get.
-// If two ReadDirRaw calls happen concurrently, the pool can hand out
-// the same buffer to both, causing data corruption.
-//
-// ReadDirRaw does: Get -> append -> return. The caller calls
-// releaseDirEntries which does Put back. If a second ReadDirRaw
-// Gets while the first caller still holds the entries, the pool
-// allocates a new one. But if the first caller releases and the
-// second Gets in sequence, the second Gets the same buffer and
-// appends over the first caller's data.
-//
-// This is actually fine in normal usage (sequential reads), but
-// there's a subtle bug: releaseDirEntries puts &s (pointer to
-// slice header with [:0]), but the underlying array still holds
-// the old data. If something retains a reference to the returned
-// entries AND another ReadDirRaw appends to the same buffer,
-// the retained reference sees corrupted data.
-
-func TestReadDirRaw_PoolAliasSafety(t *testing.T) {
-	fs, _, cleanup := newTestPassthroughFS(t)
-	defer cleanup()
-
-	// Just test that two sequential ReadDirRaw calls work correctly
-	// without pool corruption. The real issue would be if someone
-	// retains the slice across a second call.
-	entries1, err := fs.pxar.ReadDirRaw(RootInode)
-	if err != nil {
-		// No reader is set, so this returns nil entries
-		t.Logf("ReadDirRaw returned err=%v (expected, no reader)", err)
-	}
-	_ = entries1
-
-	entries2, err := fs.pxar.ReadDirRaw(RootInode)
-	if err != nil {
-		t.Logf("ReadDirRaw 2 returned err=%v", err)
-	}
-	_ = entries2
-}
-
-// Bug D: PassthroughFS.Write creates a new file handle via lazy-open
-// but doesn't record the MODIFY transaction for the backing file.
-// materializePxarFile records the MODIFY txn, but Write also does
-// lazy-open (the fh==nil path) which should also record a write txn.
-//
-// More importantly, the lazy-open path in Write doesn't call
-// fs.registerFh, it directly sets fs.handles[id] while only holding
-// fhmu. But it reads fs.nextFh under fhmu too, so that's safe.
-//
-// Actually the real issue: Write does lazy-open when fh==nil but
-// fh is 0 (the sentinel from Open). The problem is that fh==nil
-// check passes when Open returned fh=0, and then Write opens a NEW fd.
-// But if two Writes happen concurrently for the same node, both
-// see fh==nil and both open new fds — file handle leak.
-
-func TestWrite_LazyOpenHandleLeak(t *testing.T) {
-	fs, _, cleanup := newTestPassthroughFS(t)
-	defer cleanup()
-
-	fs.mutationMode = true
-
-	// Create a file with write permissions
-	var cout fuse.CreateOut
-	if st := fs.Create(nil, &fuse.CreateIn{InHeader: fuse.InHeader{NodeId: RootInode}, Mode: 0o644}, "lazyfile.txt", &cout); st != fuse.OK {
-		t.Fatalf("Create failed: %v", st)
-	}
-	ino := cout.NodeId
-
-	// Close the existing handle (simulate the fd being released)
-	fs.fhmu.Lock()
-	for id, fh := range fs.handles {
-		syscall.Close(fh.fd)
-		delete(fs.handles, id)
-	}
-	fs.fhmu.Unlock()
-
-	// Now do concurrent writes — both will see fh==nil and open new fds
-	var wg sync.WaitGroup
-	const writers = 5
-	errs := make([]fuse.Status, writers)
-
-	for i := range writers {
-		wg.Add(1)
-		go func(idx int) {
-			defer wg.Done()
-			data := []byte("hello from writer")
-			_, errs[idx] = fs.Write(nil, &fuse.WriteIn{
-				InHeader: fuse.InHeader{NodeId: ino},
-				Fh:       0, // sentinel — triggers lazy-open
-				Offset:   uint64(idx * 20),
-			}, data)
-		}(i)
-	}
-	wg.Wait()
-
-	// Count file handles — should be at most 1 per inode
-	fs.fhmu.Lock()
-	handleCount := len(fs.handles)
-	fs.fhmu.Unlock()
-
-	if handleCount > 2 { // allow some slack for race, but not 5
-		t.Errorf("expected ~1 file handle after concurrent writes, got %d (leak from lazy-open race)", handleCount)
-	}
-
-	// At least some writes should have succeeded
-	successCount := 0
-	for _, s := range errs {
-		if s == fuse.OK {
-			successCount++
-		}
-	}
-	if successCount == 0 {
-		t.Error("at least one write should have succeeded")
-	}
-}
-
-// Bug E: Fallocate has the same lazy-open race as Write.
-// Concurrent Fallocate calls for the same node with fh=0 can leak fds.
-
-func TestFallocate_LazyOpenHandleLeak(t *testing.T) {
-	fs, _, cleanup := newTestPassthroughFS(t)
-	defer cleanup()
-
-	fs.mutationMode = true
-
-	// Create a file
-	var cout fuse.CreateOut
-	if st := fs.Create(nil, &fuse.CreateIn{InHeader: fuse.InHeader{NodeId: RootInode}}, "fallocfile.txt", &cout); st != fuse.OK {
-		t.Fatalf("Create failed: %v", st)
-	}
-	ino := cout.NodeId
-
-	// Close the existing handle
-	fs.fhmu.Lock()
-	for id, fh := range fs.handles {
-		syscall.Close(fh.fd)
-		delete(fs.handles, id)
-	}
-	fs.fhmu.Unlock()
-
-	var wg sync.WaitGroup
-	const callers = 5
-	errs := make([]fuse.Status, callers)
-
-	for i := range callers {
-		wg.Add(1)
-		go func(idx int) {
-			defer wg.Done()
-			errs[idx] = fs.Fallocate(nil, &fuse.FallocateIn{
-				InHeader: fuse.InHeader{NodeId: ino},
-				Fh:       0,
-				Offset:   uint64(idx * 4096),
-				Length:   4096,
-			})
-		}(i)
-	}
-	wg.Wait()
-
-	fs.fhmu.Lock()
-	handleCount := len(fs.handles)
-	fs.fhmu.Unlock()
-
-	if handleCount > 2 {
-		t.Errorf("expected ~1 file handle after concurrent fallocate, got %d", handleCount)
-	}
-}
-
-// Bug F: metaOverlay reads are not thread-safe.
-// getOverlay reads fs.metaOverlay[relPath] without holding any lock.
-// Concurrent SetAttr can write to fs.metaOverlay via getOrCreateOverlay.
-// Test that concurrent GetAttr + SetAttr don't race on metaOverlay.
-
-func TestMetaOverlay_ConcurrentGetAttrSetAttr(t *testing.T) {
-	fs, _, cleanup := newTestPassthroughFS(t)
-	defer cleanup()
-
-	fs.mutationMode = true
-
-	// Set up root inode
-	fs.mu.Lock()
-	fs.nodePaths[RootInode] = "/"
-	fs.pathToIno["/"] = RootInode
-	fs.mu.Unlock()
-
-	// Create a non-backed pxar entry so GetAttr/SetAttr use the overlay path
-	const ino uint64 = 100 | NonDirBit
-	relPath := "/overlay-file.txt"
-	fs.pxar.mu.Lock()
-	fs.pxar.nodes[ino] = node{
-		inode:  ino,
-		parent: RootInode,
-		mode:   uint64(syscall.S_IFREG | 0o644),
-		isReg:  true,
-		refs:   1,
-	}
-	fs.pxar.mu.Unlock()
-
-	fs.mu.Lock()
-	fs.nodePaths[ino] = relPath
-	fs.pathToIno[relPath] = ino
-	// NOT backed — this forces overlay path
-	fs.mu.Unlock()
-
-	// Set up transaction log
-	mutDir := t.TempDir()
-	tl, err := OpenTransactionLog(mutDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-	fs.txnLog = tl
-
-	var wg sync.WaitGroup
-	const rounds = 50
-
-	// SetAttr writes to metaOverlay
-	wg.Go(func() {
-		for range rounds {
-			mode := uint32(0o644)
-			var sai fuse.SetAttrIn
-			sai.NodeId = ino
-			sai.Valid = fuse.FATTR_MODE
-			sai.Mode = mode
-			_ = fs.SetAttr(nil, &sai, &fuse.AttrOut{})
-		}
-	})
-
-	// GetAttr reads from metaOverlay
-	wg.Go(func() {
-		for range rounds {
-			_ = fs.GetAttr(nil, &fuse.GetAttrIn{InHeader: fuse.InHeader{NodeId: ino}}, &fuse.AttrOut{})
-		}
-	})
-
-	wg.Wait()
-}
-
-// Bug G: PassthroughFS.SetXAttr records to txnLog even when not backed
-// AND mutation mode is off — the mutationMode check is missing the
-// return before the txnLog write. Actually let me re-check...
-// SetXAttr checks !fs.mutationMode and returns EROFS, so it's fine.
-// But RemoveXAttr has the same pattern. Let me check:
-// RemoveXAttr: checks !fs.mutationMode, returns EROFS. OK fine.
-
-// Bug H: PassthroughFS.Forget deletes backed/nodePaths/pxarDir for a
-// backed node but does NOT clean up metaOverlay entries keyed by the
-// node's relPath. After Forget, stale overlay metadata persists.
-
-func TestForget_CleansUpMetaOverlay(t *testing.T) {
-	fs, _, cleanup := newTestPassthroughFS(t)
-	defer cleanup()
-
-	const ino uint64 = 100 | NonDirBit
-	path := "/forget-overlay.txt"
-
-	fs.setNode(ino, path, true)
-
-	// Add a metaOverlay entry for this path
-	fs.mu.Lock()
-	fs.metaOverlay[path] = &metaOverride{
-		mode: new(uint32(0o755)),
-		xadd: make(map[string][]byte),
-		xdel: make(map[string]bool),
-	}
-	fs.mu.Unlock()
-
-	// Forget the inode
-	fs.Forget(ino, 1)
-
-	// metaOverlay should be cleaned up for this path
-	fs.mu.RLock()
-	_, hasOverlay := fs.metaOverlay[path]
-	fs.mu.RUnlock()
-
-	if hasOverlay {
-		t.Error("metaOverlay entry should be cleaned up after Forget")
-	}
-}
-
-// Bug I: PassthroughFS.Lookup does backedNode.refs++ without holding mu
-// in the "backed directory that's also a pxar dir" path.
-// Line: fs.mu.Lock(); backedNode.refs++; backedNode.parent = header.NodeId; fs.mu.Unlock()
-// Actually it does hold mu. Let me check another issue...
-
-// Bug J: PassthroughFS.Lookup reads fs.metaOverlay[childPath] without
-// holding mu. The rename-from overlay check:
-//   if mo := fs.getOverlay(childPath); mo != nil && mo.renameFrom != "" {
-// getOverlay reads fs.metaOverlay[childPath] without any lock.
-
-// This is already covered by Bug F's pattern. Let me check getOverlay:
-// func (fs *PassthroughFS) getOverlay(relPath string) *metaOverride {
-//     return fs.metaOverlay[relPath]
-// }
-// No lock! This is a data race.
-
-// Bug K: Concurrent renamePxarEntry cleans up overlays by iterating
-// fs.metaOverlay without holding a lock consistently.
-// The loop:
-//   for path, mo := range fs.metaOverlay {
-//       if mo.renameFrom == oldPath { delete(fs.metaOverlay, path) }
-//   }
-// is inside renamePxarEntry which doesn't hold fs.mu for this iteration.
-
-func TestRenamePxarEntry_OverlayCleanupRace(t *testing.T) {
-	fs, _, cleanup := newTestPassthroughFS(t)
-	defer cleanup()
-
-	fs.mutationMode = true
-
-	// Set up root
-	fs.mu.Lock()
-	fs.nodePaths[RootInode] = "/"
-	fs.pathToIno["/"] = RootInode
-	fs.mu.Unlock()
-
-	// Create some overlay entries that will be iterated
-	for i := range 10 {
-		path := "/renamed-file-" + string(rune('a'+i)) + ".txt"
-		fs.mu.Lock()
-		fs.metaOverlay[path] = &metaOverride{
-			renameFrom: "/src-dir/old-file.txt",
-			xadd:       make(map[string][]byte),
-			xdel:       make(map[string]bool),
-		}
-		fs.mu.Unlock()
-	}
-
-	var wg sync.WaitGroup
-	const rounds = 20
-
-	// Writer: simulates renamePxarEntry iterating metaOverlay
-	wg.Go(func() {
-		for range rounds {
-			fs.mu.Lock()
-			for path, mo := range fs.metaOverlay {
-				if mo.renameFrom == "/src-dir/old-file.txt" {
-					delete(fs.metaOverlay, path)
-				}
-			}
-			fs.mu.Unlock()
-
-			// Re-add entries for next round
-			fs.mu.Lock()
-			for i := range 10 {
-				path := "/renamed-file-" + string(rune('a'+i)) + ".txt"
-				fs.metaOverlay[path] = &metaOverride{
-					renameFrom: "/src-dir/old-file.txt",
-					xadd:       make(map[string][]byte),
-					xdel:       make(map[string]bool),
-				}
-			}
-			fs.mu.Unlock()
-		}
-	})
-
-	// Reader: GetAttr reads metaOverlay via getOverlay (no lock)
-	wg.Go(func() {
-		for range rounds {
-			_ = fs.getOverlay("/renamed-file-a.txt")
-		}
-	})
-
-	wg.Wait()
-	// This will flag with -race due to getOverlay accessing map without lock
-}
